Splits Autocomplete from Search, and fixes functionality
